requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-multipart
python-jose
httpx
authlib
itsdangerous
databases
aiosqlite
python-dotenv
